feat: namespace mux clients

+++ cmd/skirk/parent_watch_android.go
@@ -5,11 +5,14 @@ package main
 import (
 	"context"
 	"log"
+	"os"
 	"syscall"
 	"time"
 )
 
-func enableParentDeathSignal() {}
+func enableParentDeathSignal() {
+	_, _, _ = syscall.RawSyscall(syscall.SYS_PRCTL, 1, uintptr(syscall.SIGTERM), 0)
+}
 
 func watchParentProcess(ctx context.Context, pid int, cancel context.CancelFunc) {
 	go func() {
@@ -20,6 +23,11 @@ func watchParentProcess(ctx context.Context, pid int, cancel context.CancelFunc)
 			case <-ctx.Done():
 				return
 			case <-ticker.C:
+				if os.Getppid() == 1 {
+					log.Printf("parent process disappeared pid=%d", pid)
+					cancel()
+					return
+				}
 				if err := syscall.Kill(pid, 0); err == syscall.ESRCH {
 					log.Printf("parent process disappeared pid=%d", pid)
 					cancel()
